refactor(ts-sdk): split streaming-backend tests by function (S4-07)
Devin finding 4 on PR #595: streaming-backend.test.ts was 616 lines,
exceeding the 500-line file NLOC limit. Split by function under test:

  tests/streaming-backend.test.ts          (236 lines) — streamUpsertPoints
  tests/streaming-backend-train-pq.test.ts (118 lines) — trainPq
  tests/streaming-backend-insert.test.ts   (262 lines) — streamInsert
  tests/helpers/build-streaming-transport.ts (52 lines) — shared factory

The buildTransport() factory is extracted into a shared helper to avoid
triplicating ~25 lines across the split files; this keeps each test
file self-contained (tests directly import from a single helper module)
without DRY violations flagged by jscpd.

All 32 streaming-backend tests preserved (12 + 7 + 13 = 32). Full test
suite (401 tests) passes, coverage thresholds met.
